sacct service added and separated wm helpers from slurm.py
